Fix signed-to-unsigned cast w/ overflow potential
Discovered by Yves Younan of Cisco Talos and assigned
tracking ID TALOS-CAN-0064.

+++ include/ntp_refclock.h
@@ -99,7 +99,7 @@ struct refclockio {
 				due to small bursts
 				of refclock input data */
 	struct peer *srcclock;	/* refclock peer */
-	u_int	datalen;	/* length of data */
+	size_t	datalen;	/* length of data */
 	int	fd;		/* file descriptor */
 	u_long	recvcount;	/* count of receive completions */
 	bool	active;		/* true when in use */
